backend/pyproject.toml
[project]
name = "figma-to-html-backend"
version = "0.1.0"
description = "Figma-to-HTML/CSS converter backend with AI-powered pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "websockets>=12.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "openai>=1.30.0",
    "playwright>=1.44.0",
    "Pillow>=10.3.0",
    "scikit-image>=0.23.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
